+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -35,6 +35,7 @@ however, it has to be formatted properly to pass verification tests.
 - Fixed lingering highlight effect on Save Asset button after clicking.
 - Fixed missing name in window title for Input Action assets.
 - Fixed "Listen" functionality for selecting an input sometimes expecting the wrong input type.
+- Fixed console errors that can be produced when opening input package settings from the Inspector.
 
 ## [1.8.0-pre.2] - 2023-11-09